+++ Example/Example/Data/MatchRepository.swift
@@ -12,21 +12,24 @@ class MatchRepository {
     func loadMatches() -> AnyPublisher<[MatchMarkets], Never> {
         let matches = MatchMarkets.simulatedMatches
 
-        return identityMap.store(matches, modifiedAt: MatchMarkets.simulatedFetchedDate.stamp)
+        return identityMap.store(
+            matches,
+            using: Relations.matchMarkets,
+            modifiedAt: MatchMarkets.simulatedFetchedDate.stamp
+        )
     }
 
     /// observe primary (first) match market changes (for this sample changes are generated randomely
     /// - Returns: the match with all its markets including updates for primary market
     func observePrimaryMarket(for match: Match) -> AnyPublisher<MatchMarkets, Never> {
         let data = MatchMarkets.simulatedMatches.first { $0.match.id == match.id }!
+        let outcomes = data.primaryMarket.outcomes.map { identityMap.get(for: Outcome.self, id: $0.id) ?? $0 }
 
-        return data
-            .primaryMarket
-            .outcomes
+        return outcomes
             .map { outcome in self.randomChanges(for: outcome) }
             .combineLatest()
             .map { [identityMap] in identityMap.store($0) }
-            .map { [identityMap] _ in identityMap.publisher(for: MatchMarkets.self, id: match.id) }
+            .map { [identityMap] _ in identityMap.publisher(using: Relations.matchMarkets, id: match.id) }
             .switchToLatest()
             .eraseToAnyPublisher()
     }