+++ Sources/CohesionKit/Storage.swift
@@ -9,40 +9,24 @@ import Foundation
 import Combine
 import CombineExt
 
-struct StampedObject<D> {
-    let object: D
-    let lastModification: Stamp
-}
+typealias StampedObject<D> = (object: D, modifiedAt: Stamp)
 
 class Storage<T> {
-    // TODO: Set as private
-    let subject: CurrentValueSubject<StampedObject<T>?, Never>
+    private let subject: CurrentValueSubject<StampedObject<T>?, Never>
     private(set) var publisher: AnyPublisher<T, Never>!
     private var upstreamCancellable: AnyCancellable?
+    
+    var value: T? { subject.value?.object }
 
-    /// init storage with a initial value for a `Idenfitiable` object
-    convenience init(object: T, modifiedAt: Stamp, identityMap: IdentityMap) where T: Identifiable {
-        self.init(StampedObject(object: object, lastModification: modifiedAt)) { [weak identityMap] in
-            identityMap?.remove(object)
+    /// init an empty storage
+    convenience init(id: Any, identityMap: IdentityMap) {
+        self.init() { [weak identityMap] in
+            identityMap?[T.self, id: id] = nil
         }
     }
 
-    /// init an empty storage for a `Idenfitiable` object
-    convenience init(id: T.ID, identityMap: IdentityMap) where T: Identifiable {
-        self.init(nil) { [weak identityMap] in
-            identityMap?.remove(for: T.self, id: id)
-        }
-    }
-
-    /// init an empty storage for an `IdentityGraph` object using its id
-    convenience init(id: T.Identity.ID, identityMap: IdentityMap) where T: Relational {
-        self.init(nil) { [weak identityMap] in
-            identityMap?.remove(for: T.self, id: id)
-        }
-    }
-
-    private init(_ object: StampedObject<T>?, remove: @escaping () -> Void) {
-        self.subject = CurrentValueSubject(object)
+    init(remove: @escaping () -> Void) {
+        self.subject = CurrentValueSubject(nil)
         self.publisher = subject
             .compactMap { $0?.object }
             .handleEvents(receiveCancel: { [weak self] in
@@ -59,16 +43,16 @@ class Storage<T> {
     /// - Returns: true if storage was updated. Storage is updated only if `stamp` is sup. to storage stamp
     @discardableResult
     func send(_ input: T, modifiedAt: Stamp) -> Bool {
-        guard subject.value.map({ modifiedAt >= $0.lastModification }) ?? true else {
+        guard subject.value.map({ modifiedAt >= $0.modifiedAt }) ?? true else {
             return false
         }
 
-        subject.send(StampedObject(object: input, lastModification: modifiedAt))
+        subject.send((object: input, modifiedAt: modifiedAt))
         return true
     }
 
-    /// Forward a `Publisher` and send its value to Storage
-    func forward(_ upstream: AnyPublisher<T, Never>, modifiedAt: Stamp) {
+    /// Merge value from `upstream` into the storage
+    func merge(_ upstream: AnyPublisher<T, Never>, modifiedAt: Stamp) {
         upstreamCancellable = upstream
             .sink(receiveValue: { [weak self] in self?.send($0, modifiedAt: modifiedAt) })
     }